Remove legacy navigation code from the project
