+++ apps/browser/src/main/res/layout/activity_main.xml
@@ -1,17 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.coordinatorlayout.widget.CoordinatorLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:context=".MainActivity">
-
-  <com.bluelinelabs.conductor.ChangeHandlerFrameLayout
-      android:id="@+id/controller_container"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent" />
 
   <androidx.fragment.app.FragmentContainerView
       android:id="@+id/nav_host_fragment"
