+++ features/projects/impl/src/main/java/com/kuba/example/projects/impl/search/RepositorySearchViewModel.kt
@@ -1,19 +1,19 @@
 package com.kuba.example.projects.impl.search
 
 import androidx.lifecycle.ViewModel
-import com.kuba.example.dagger.conductor.scope.ControllerScope
-import com.kuba.example.projects.impl.contributors.ContributorsController
 import com.kuba.example.service.api.GithubService
 import com.kuba.example.service.api.ServiceResult
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
 
-@ControllerScope(RepositorySearchController::class)
+@HiltViewModel
 class RepositorySearchViewModel @Inject constructor(private val githubService: GithubService) :
     ViewModel() {
 
-    private val _state: MutableStateFlow<RepositoriesUiModel> = MutableStateFlow(RepositoriesUiModel.Content(emptyList()))
+    private val _state: MutableStateFlow<RepositoriesUiModel> =
+        MutableStateFlow(RepositoriesUiModel.Content(emptyList()))
     val state: StateFlow<RepositoriesUiModel> = _state
 
     /**
