+++ features/users/impl/src/main/java/com/kuba/example/users/impl/repos/UserRepositoriesViewModel.kt
@@ -1,24 +1,26 @@
 package com.kuba.example.users.impl.repos
 
+import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import com.kuba.example.dagger.conductor.scope.ControllerScope
 import com.kuba.example.service.api.GithubService
 import com.kuba.example.service.api.ServiceResult
-import com.kuba.example.service.api.User
+import com.kuba.example.users.api.navigation.UserRepositoriesScreen
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import javax.inject.Inject
 
-@ControllerScope(UserRepositoriesController::class)
+@HiltViewModel
 class UserRepositoriesViewModel @Inject constructor(
-    private val user: User, // example of injecting controller-scoped dependencies
-    private val githubService: GithubService
-) :
-    ViewModel() {
+    private val githubService: GithubService,
+    private val savedStateHandle: SavedStateHandle
+) : ViewModel() {
 
     private val _state: MutableStateFlow<UserRepositoriesUiModel> =
         MutableStateFlow(UserRepositoriesUiModel.Content(emptyList()))
     val state: StateFlow<UserRepositoriesUiModel> = _state
+
+    private val user get() = UserRepositoriesScreen.extractUser(savedStateHandle)
 
     /**
      * Search Github repositories
