+++ settings.gradle
@@ -31,4 +31,3 @@ include ':shared:service:api'
 include ':shared:service:impl'
 include ':shared:navigation:api'
 include ':shared:navigation:impl'
-include ':shared:dagger-conductor'
