--- shared/dagger-conductor/build.gradle
@@ -1,46 +1,0 @@
-plugins {
-    alias(libs.plugins.androidLibrary)
-    alias(libs.plugins.jetbrainsKotlinAndroid)
-    alias(libs.plugins.ksp)
-}
-
-android {
-    namespace 'com.kuba.example.dagger.conductor'
-    compileSdk 34
-
-    defaultConfig {
-        minSdk 32
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-}
-
-dependencies {
-    // Core
-    implementation libs.timber
-
-    // DI
-    implementation libs.dagger.hilt.android
-    ksp libs.dagger.hilt.compiler
-    implementation libs.dagger.android
-
-    // Navigation
-    implementation libs.conductor
-    implementation libs.androidx.core.ktx
-    implementation libs.androidx.appcompat
-}
