--- shared/navigation/api/src/main/java/com/kuba/example/navigation/api/ControllerFactory.kt
@@ -1,11 +1,0 @@
-package com.kuba.example.navigation.api
-
-import com.bluelinelabs.conductor.Controller
-
-/**
- * Creates real [Controller]s from [ControllerDestination] specs
- */
-interface ControllerFactory {
-
-    fun  <T : ControllerDestination> create(controllerDestination: T) : Controller
-}
