--- shared/navigation/impl/src/main/java/com/kuba/example/navigation/di/ControllerBuilders.kt
@@ -1,7 +1,0 @@
-package com.kuba.example.navigation.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.RUNTIME)
-annotation class ControllerBuilders
