--- shared/navigation/impl/src/main/java/com/kuba/example/navigation/di/NavModule.kt
@@ -1,23 +1,0 @@
-package com.kuba.example.navigation.di
-
-import com.kuba.example.navigation.MultiBoundControllerFactory
-import com.kuba.example.navigation.api.ControllerBuilder
-import com.kuba.example.navigation.api.ControllerDestination
-import com.kuba.example.navigation.api.ControllerFactory
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import dagger.multibindings.Multibinds
-
-@InstallIn(SingletonComponent::class)
-@Module
-interface NavModule {
-
-    @Binds
-    fun controllerDestinations(multiBoundControllerFactory: MultiBoundControllerFactory): ControllerFactory
-
-    @ControllerBuilders
-    @Multibinds
-    fun controllerBuilders(): Map<Class<out ControllerDestination>, ControllerBuilder>
-}
